requirements.txt
# Scraping
requests>=2.32,<3
beautifulsoup4>=4.12,<5
lxml>=5.3

# Datos
pandas>=2.2
python-dateutil>=2.9

# Base de datos en la nube (Supabase)
# Nota: fijamos <2.22 porque a partir de esa version storage3 arrastra pyiceberg,
# que no tiene wheels para Python 3.14 y requiere MSVC para compilarse.
# La API del cliente es la misma en este rango.
supabase>=2.15,<2.22
python-dotenv>=1.0

# Dashboard web
streamlit>=1.39
plotly>=5.24

# Utilidades
tqdm>=4.66
